+++ src/Http/Middleware/Pipeline/HandleRedirectResponse.php
@@ -19,17 +19,11 @@ use Hamzi\Catchy\Support\FlashExtractor;
 class HandleRedirectResponse implements PipelineStageInterface
 {
     /**
-     * The asset version repository instance.
-     */
-    protected VersionRepositoryInterface $versionRepository;
-
-    /**
      * HandleRedirectResponse constructor.
      */
-    public function __construct(VersionRepositoryInterface $versionRepository)
-    {
-        $this->versionRepository = $versionRepository;
-    }
+    public function __construct(
+        protected readonly VersionRepositoryInterface $versionRepository
+    ) {}
 
     /**
      * Handle the pipeline stage.